+++ CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/) and this
 project adheres to [Semantic Versioning](https://semver.org/).
 
+# 0.3.1 - 2021-11-08
+
+## Added
+
+## Fixed
+
+- Grouping multiple X-properties of the same type into one group resulted in
+  lost properties. Found while tuning code coverage.
+
+## Changed
+
 # 0.3.0 - 2021-11-07
 
 ## Added